Guard SourceExt::write against no-op resubmits, split text widgets out
